requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-excel==2.4.0
drf-renderer-xlsx==1.0.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.1
pipenv==2023.10.3
platformdirs==3.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
XlsxWriter==3.1.6
xlwt==1.3.0
